LogBox and Toastify. toast index
(cherry picked from commit 5e8abd612c78fe08de740d0a4ed858b9219b15ef)
