+++ front-end-mobile/src/shared/components/toast/Toast.tsx
@@ -8,31 +8,22 @@ import {
   SafeAreaView,
   View,
   StyleSheet,
-  Text,
-  Image,
   Platform,
-  TouchableOpacity,
 } from 'react-native';
-import R from 'shared/res/R';
-import FadingView, {
+import {
   FadingViewRef,
 } from '../fadingView/FadingView';
+import {
+  ToastType,
+  ToastTheme,
+  ToastMessage,
+  ToastRef,
+} from './types';
+import Toastify from './Toast.Toastify';
+import LogBox from './Toast.LogBox';
 
-type ToastType = 'error' | 'info' | 'warn' | 'success' | 'log';
-
-export type ToastMessage = {
-  type: ToastType,
-  content: string;
-}
-export const isToastMessageEqual = (message1: ToastMessage, message2: ToastMessage) => {
-  return message1.content === message2.content && message1.type === message2.type;
-};
-
-type ToastProps = {}
-
-export type ToastRef = {
-  show: (toastMsg: ToastMessage, onClose?: () => void) => void;
-  hide: () => void;
+type ToastProps = {
+  theme?: ToastTheme;
 }
 
 type PropsWithForwardedRef = ToastProps & {
@@ -79,23 +70,38 @@ const Toast: React.FC<PropsWithForwardedRef> = (props: PropsWithForwardedRef) =>
     }
   }, [isShowing, _onClose]);
 
+  const _renderToast = (theme?: ToastTheme) => {
+    theme = theme ? theme : 'Toastify';
+
+    switch (theme) {
+      case 'Toastify':
+        return (
+          <Toastify
+            type={type}
+            message={message}
+            onClose={_hidePopup}
+            ref={fadingViewRef}
+          />
+        );
+
+      case 'LogBox':
+      default:
+        return (
+          <LogBox
+            type={type}
+            message={message}
+            onClose={_hidePopup}
+            ref={fadingViewRef}
+          />
+        );
+    }
+  };
+
   return (
     <>
       {isShowing && <View style={styles.absoluteContainer}>
         <SafeAreaView style={styles.safeAreaContainer}>
-          <FadingView style={styles.container(type)} ref={fadingViewRef}>
-            <View style={styles.textContainer}>
-              <Text style={styles.text(type)}>{message}</Text>
-            </View>
-            <View style={styles.closeButtonContainer}>
-              {/* If you cannot click on the button, make sure you put the Toast under other views*/}
-              <TouchableOpacity
-                style={styles.closeButton}
-                onPress={_hidePopup}>
-                <Image source={R.images.ic_close} style={styles.closeImage(type)} />
-              </TouchableOpacity>
-            </View>
-          </FadingView>
+          {_renderToast(props.theme)}
         </SafeAreaView>
       </View>}
     </>
@@ -104,91 +110,23 @@ const Toast: React.FC<PropsWithForwardedRef> = (props: PropsWithForwardedRef) =>
 
 const padding = 20;
 const safeAreaContainerPadding = 10;
-const closeButtonTop = 10;
-const closeButtonPadding = 5;
 
-const shared = (type?: ToastType) => {
-  let bgColor: string = R.colors.BLACK;
-  let contentColor: string = R.colors.WHITE;
-
-  switch (type) {
-    case 'error':
-      bgColor = R.colors.SUNSET;
-      break;
-
-    case 'info':
-      bgColor = R.colors.BLUE;
-      break;
-
-    case 'success':
-      bgColor = R.colors.GREEN;
-      break;
-
-    case 'warn':
-      bgColor = R.colors.YELLOW;
-      contentColor = R.colors.BLACK;
-      break;
-
-    case 'log':
-      R.colors.BLACK;
-      break;
-  }
-
-  return StyleSheet.create({
-    absoluteContainer: {
-      position: 'absolute',
-      width: '100%',
-      alignSelf: 'center',
-      paddingHorizontal: padding - safeAreaContainerPadding,
-      marginTop: Platform.select({
-        ios: 20,
-        android: 0,
-      }),
-    },
-    safeAreaContainer: {
-      flex: 1,
-      padding: safeAreaContainerPadding,
-    },
-    container: {
-      alignItems: 'center',
-      backgroundColor: bgColor,
-      borderRadius: 5,
-      elevation: 4,
-    },
-    closeButtonContainer: {
-      position: 'absolute',
-      top: closeButtonTop - closeButtonPadding,
-      right: closeButtonTop - closeButtonPadding,
-    },
-    closeImage: {
-      tintColor: contentColor,
-    },
-    closeButton: {
-      padding: closeButtonPadding,
-    },
-    textContainer: {
-      flex: 1,
-      width: '100%',
-      padding: 15,
-    },
-    text: {
-      ...R.palette.normal,
-      color: contentColor,
-      textAlign: 'center',
-    },
-  });
-};
-
-const styles = {
-  absoluteContainer: shared().absoluteContainer,
-  safeAreaContainer: shared().safeAreaContainer,
-  container: (type: ToastType) => shared(type).container,
-  closeButtonContainer: shared().closeButtonContainer,
-  closeImage: (type: ToastType) => shared(type).closeImage,
-  closeButton: shared().closeButton,
-  textContainer: shared().textContainer,
-  text: (type: ToastType) => shared(type).text,
-};
+const styles = StyleSheet.create({
+  absoluteContainer: {
+    position: 'absolute',
+    width: '100%',
+    alignSelf: 'center',
+    paddingHorizontal: padding - safeAreaContainerPadding,
+    marginTop: Platform.select({
+      ios: 20,
+      android: 0,
+    }),
+  },
+  safeAreaContainer: {
+    flex: 1,
+    padding: safeAreaContainerPadding,
+  },
+});
 
 // eslint-disable-next-line react/display-name
 export default React.forwardRef((props: ToastProps, ref: React.Ref<ToastRef>) => <Toast {...props} myForwardedRef={ref} />);
